MM-90: migrate public providers routes to experts
