+++ marketlink-frontend/src/app/experts/loading.tsx
@@ -1,0 +1,1 @@
+export { default } from '../providers/loading';
